requirements.txt
fastapi
uvicorn
httpx
python-multipart
google-cloud-storage==2.10.0